# backend/openapi.yaml
openapi: 3.1.0
info:
  title: SNT Portal API
  version: 1.0.0
  description: |
    Unified API for resident/chairman portal.
servers:
  - url: http://localhost:3000
    description: local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/auth/change-password:
    post:
      summary: Change password for current user
      security:
        - bearerAuth: []
  /api/v1/auth/logout:
    post:
      summary: Revoke session
      security:
        - bearerAuth: []
...
  /api/v1/chat/rooms:
    get:
      summary: List chat rooms
      security:
        - bearerAuth: []
    post:
      summary: Create chat room (chairman)
      security:
        - bearerAuth: []
...